+++ src/apis/brand/BrandDto.ts
@@ -1,3 +1,11 @@
+export interface CreateBrandRequest {
+  name: string
+}
+
+export interface UpdateBrandRequest {
+  name: string
+}
+
 export interface ReadBrandResponse {
   id: number
   name: string
