--- src/apis/brand/dto/BrandRequest.ts
@@ -1,7 +1,0 @@
-export interface CreateBrandRequest {
-  name: string
-}
-
-export interface UpdateBrandRequest {
-  name: string
-}
